+++ asciimatics/widgets/textbox.py
@@ -274,7 +274,7 @@ class TextBox(Widget):
                     column = 0
                     while self.string_len(str(line)) >= limit:
                         sub_string = _enforce_width(
-                            line, limit, self._frame.canvas.unicode_aware)
+                            line, limit, self._frame.canvas.unicode_aware, split_on_words=True)
                         self._reflowed_text_cache.append((sub_string, i, column))
                         line = line[len(sub_string):]
                         column += len(sub_string)